python/requirements.txt
flask==1.0
requests==2.19.1
pycrypto==2.6.1
